% Generated by roxygen2: do not edit by hand
% Please edit documentation in ./import_module.R
\name{import_module_help}
\alias{import_module_help}
\title{Open documentation for \code{import_module}}
\usage{
import_module_help()
}
\description{
Open a rendered HTML page of \code{import_module} documentation.
See \url{https://github.com/bobaekang/r-module-trick} for a detailed explanation.
}
