% Generated by roxygen2: do not edit by hand
% Please edit documentation in ./import_module.R
\name{import_module}
\alias{import_module}
\title{Import an R "module"}
\usage{
import_module(path, name, attached = TRUE)
}
\arguments{
\item{path}{A character string for the path to a module file.}

\item{name}{A character string for an optional module name.}

\item{attached}{A logical value. If TRUE, module is attached to the search
path. If FALSE, a module object is created in the global environment.}
}
\description{
Import a "module" from \code{path} using an optional \code{name}.
See \url{https://github.com/bobaekang/r-module-trick} for a detailed explanation.
}
\examples{
# import a local module file
import_module(path = "module.R")

# import a remote module file with an optional name
path <- "https://tinyurl.com/r-module-trick/module.R"
import_module(path = path, name = "module_example")
}
\seealso{
\code{\link[base]{attach}} for attaching R object to search path.

\code{\link[base]{assign}} for assigning a value to name.
}
